+++ src/remarks/remarks.service.ts
@@ -1,10 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Remark } from './remark.entity';
 import { CreateRemarkDto } from './dto/create-remark.dto';
 import { UsersService } from '../users/users.service';
 import { Post } from '../posts/post.entity';
+import { RemarkNotFoundByIdException } from './exception/remark-not-found-by-id-exception';
 
 @Injectable()
 export class RemarksService {
@@ -15,16 +21,17 @@ export class RemarksService {
   ) {}
 
   async createRemark(remarkCreate: CreateRemarkDto): Promise<Remark> {
-    const posts: Post[] = await this.postRepository.find({
-      where: { id: remarkCreate.idPost },
-    });
-    const remark: Remark = this.remarkRepository.create({
-      post: posts[0],
-      idParentRemark: remarkCreate.idParentRemark,
-      content: remarkCreate.content,
-      user: await this.usersService.findByUserId(remarkCreate.idUser),
-    });
-    return this.remarkRepository.save(remark);
+    try {
+      const remark: Remark = this.remarkRepository.create({
+        post: await this.postRepository.findOne(remarkCreate.idPost),
+        idParentRemark: remarkCreate.idParentRemark,
+        content: remarkCreate.content,
+        user: await this.usersService.findByUserId(remarkCreate.idUser),
+      });
+      return this.remarkRepository.save(remark);
+    } catch (error) {
+      throw new BadRequestException(remarkCreate, 'Follow creation error');
+    }
   }
 
   async getAll(): Promise<Remark[]> {
@@ -32,10 +39,11 @@ export class RemarksService {
   }
 
   async findByRemarkId(remarkId: number): Promise<Remark> {
-    const remarks = await this.remarkRepository.find({
-      where: { id: remarkId },
-    });
-    return remarks[0];
+    const remark = await this.remarkRepository.findOne(remarkId);
+    if (remark) {
+      return remark;
+    }
+    throw new RemarkNotFoundByIdException(remarkId);
   }
 
   async findByPostId(postId: number): Promise<Remark[]> {
@@ -45,7 +53,11 @@ export class RemarksService {
   }
 
   async deleteRemarksById(id: number) {
-    return await this.remarkRepository.delete(id);
+    const remark = await this.remarkRepository.delete(id);
+    if (remark) {
+      return await this.remarkRepository.delete(id);
+    }
+    throw new RemarkNotFoundByIdException(id);
   }
 
   async deleteRemarksByPostId(postId: number) {
@@ -55,7 +67,10 @@ export class RemarksService {
         this.deleteRemarksById(remark.id);
       });
     } catch (error) {
-      console.log('likes delete error: ' + error);
+      throw new HttpException(
+        'Remarks delete error: ' + error,
+        HttpStatus.BAD_REQUEST,
+      );
     }
   }
 }
